src/requirements/requirements.txt
alembic==1.0.8
asn1crypto==0.24.0
bcrypt==3.1.6
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
coverage==4.5.3
cryptography==2.6.1
Fabric3==1.14.post1
Flask==1.0.2
Flask-JSON==0.3.3
Flask-Migrate==2.4.0
flask-peewee==3.0.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.3.2
Mako==1.0.8
MarkupSafe==1.1.1
marshmallow==2.19.1
Marshmallow-Peewee==2.2.0
paramiko==2.4.2
peewee==3.9.2
pyasn1==0.4.5
pycparser==2.19
pycrypto==2.6.1
PyMySQL==0.9.3
PyNaCl==1.3.0
python-dateutil==2.8.0
python-editor==1.0.4
requests==2.21.0
six==1.12.0
soupsieve==1.8
SQLAlchemy==1.3.1
urllib3==1.24.1
Werkzeug==0.15.1
wtf-peewee==3.0.0
WTForms==2.2.1
